调整节点算法代码的位置
